Move shader proto to its own proto file under element package
